[BACKLOG-21825] Remove pentaho.type.{Model, Application}
