--- impl/client/src/main/doc-js/pentaho/type/spec/IModel.jsdoc
@@ -1,34 +1,0 @@
-/*!
- * Copyright 2010 - 2017 Hitachi Vantara. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * The `spec.IModel` interface describes a model instance in object form.
- *
- * @name pentaho.type.spec.IModel
- *
- * @interface
- * @extends pentaho.type.spec.IComplex
- *
- * @see pentaho.type.spec.UComplex
- */
-
-/**
- * Represents the relevant state and interface of the application in which a model is being used.
- *
- * @name application
- * @memberOf pentaho.type.spec.IModel#
- * @type {pentaho.type.spec.IApplication|pentaho.type.Application}
- */
