+++ impl/client/src/main/javascript/web/pentaho/type/i18n/types.properties
@@ -42,10 +42,6 @@ typeDescriptor.description=A type descriptor value.
 ## MODEL
 model.label=Model
 model.description=A model of an application.
-## APPLICATION
-application.label=Application
-application.description=An application.
-
 ## ACTION
 action.label=Action
 action.description=An action.
